Move Curriculum to domain
